refactored files and created new directories
